requirements.txt
streamlit
langchain
chromadb
requests
python-dotenv
google-generativeai
pdfplumber
langchain-community